feat: added ContractService class
