+++ packages/relay/src/lib/services/ethService/ethCommonService/index.ts
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { ConfigService } from '@hashgraph/json-rpc-config-service/dist/services';
+import { Hbar } from '@hashgraph/sdk';
 import * as _ from 'lodash';
 import { Logger } from 'pino';
 
@@ -15,7 +16,9 @@ import {
   stripLeadingZeroForSignatures,
   tinybarsToWeibars,
   toHash32,
+  weibarHexToTinyBarInt,
 } from '../../../../formatters';
+import { Utils } from '../../../../utils';
 import { MirrorNodeClient } from '../../../clients';
 import constants from '../../../constants';
 import { JsonRpcError, predefined } from '../../../errors/JsonRpcError';
@@ -23,13 +26,11 @@ import { MirrorNodeClientError } from '../../../errors/MirrorNodeClientError';
 import { SDKClientError } from '../../../errors/SDKClientError';
 import { EthImpl } from '../../../eth';
 import { Log } from '../../../model';
-import { RequestDetails } from '../../../types';
+import { IAccountInfo, IContractCallRequest, RequestDetails } from '../../../types';
 import { CacheService } from '../../cacheService/cacheService';
-import { ICommonService } from './ICommonService';
-import { Utils } from '../../../../utils';
-import { Hbar } from '@hashgraph/sdk';
-import HAPIService from '../../hapiService/hapiService';
 import { TransactionFactory } from '../../factories/transactionFactory';
+import HAPIService from '../../hapiService/hapiService';
+import { ICommonService } from './ICommonService';
 
 /**
  * Create a new Common Service implementation.
@@ -79,11 +80,11 @@ export class CommonService implements ICommonService {
   private readonly maxBlockRange = parseNumericEnvVar('MAX_BLOCK_RANGE', 'MAX_BLOCK_RANGE');
   private readonly ethBlockNumberCacheTtlMs = parseNumericEnvVar(
     'ETH_BLOCK_NUMBER_CACHE_TTL_MS',
-    'ETH_BLOCK_NUMBER_CACHE_TTL_MS_DEFAULT'
+    'ETH_BLOCK_NUMBER_CACHE_TTL_MS_DEFAULT',
   );
   private readonly ethGasPriceCacheTtlMs = parseNumericEnvVar(
     'ETH_GET_GAS_PRICE_CACHE_TTL_MS',
-    'ETH_GET_GAS_PRICE_CACHE_TTL_MS_DEFAULT'
+    'ETH_GET_GAS_PRICE_CACHE_TTL_MS_DEFAULT',
   );
 
   // Maximum allowed timestamp range for mirror node requests' timestamp parameter is 7 days (604800 seconds)
@@ -93,7 +94,12 @@ export class CommonService implements ICommonService {
     return ConfigService.get('ETH_GET_LOGS_BLOCK_RANGE_LIMIT');
   }
 
-  constructor(mirrorNodeClient: MirrorNodeClient, logger: Logger, cacheService: CacheService, hapiService: HAPIService) {
+  constructor(
+    mirrorNodeClient: MirrorNodeClient,
+    logger: Logger,
+    cacheService: CacheService,
+    hapiService: HAPIService,
+  ) {
     this.mirrorNodeClient = mirrorNodeClient;
     this.logger = logger;
     this.cacheService = cacheService;
@@ -119,7 +125,7 @@ export class CommonService implements ICommonService {
     fromBlock: string,
     toBlock: string,
     requestDetails: RequestDetails,
-    address?: string | string[] | null
+    address?: string | string[] | null,
   ) {
     if (this.blockTagIsLatestOrPending(toBlock)) {
       toBlock = CommonService.blockLatest;
@@ -177,7 +183,7 @@ export class CommonService implements ICommonService {
           prepend0x(fromBlockNum.toString(16)),
           fromBlockResponse.timestamp.from,
           prepend0x(toBlockNum.toString(16)),
-          toBlockResponse.timestamp.to
+          toBlockResponse.timestamp.to,
         );
       }
 
@@ -262,7 +268,7 @@ export class CommonService implements ICommonService {
     if (!returnLatest && this.blockTagIsLatestOrPending(blockNumberOrTagOrHash)) {
       if (this.logger.isLevelEnabled('debug')) {
         this.logger.debug(
-          `${requestDetails.formattedRequestId} Detected a contradiction between blockNumberOrTagOrHash and returnLatest. The request does not target the latest block, yet blockNumberOrTagOrHash representing latest or pending: returnLatest=${returnLatest}, blockNumberOrTagOrHash=${blockNumberOrTagOrHash}`
+          `${requestDetails.formattedRequestId} Detected a contradiction between blockNumberOrTagOrHash and returnLatest. The request does not target the latest block, yet blockNumberOrTagOrHash representing latest or pending: returnLatest=${returnLatest}, blockNumberOrTagOrHash=${blockNumberOrTagOrHash}`,
         );
       }
       return null;
@@ -271,7 +277,7 @@ export class CommonService implements ICommonService {
     if (blockNumberOrTagOrHash === EthImpl.emptyHex) {
       if (this.logger.isLevelEnabled('debug')) {
         this.logger.debug(
-          `${requestDetails.formattedRequestId} Invalid input detected in getHistoricalBlockResponse(): blockNumberOrTagOrHash=${blockNumberOrTagOrHash}.`
+          `${requestDetails.formattedRequestId} Invalid input detected in getHistoricalBlockResponse(): blockNumberOrTagOrHash=${blockNumberOrTagOrHash}.`,
         );
       }
       return null;
@@ -311,15 +317,15 @@ export class CommonService implements ICommonService {
     const blockNumberCached = await this.cacheService.getAsync(
       cacheKey,
       CommonService.latestBlockNumber,
-      requestDetails
+      requestDetails,
     );
 
     if (blockNumberCached) {
       if (this.logger.isLevelEnabled('trace')) {
         this.logger.trace(
           `${requestDetails.formattedRequestId} returning cached value ${cacheKey}:${JSON.stringify(
-            blockNumberCached
-          )}`
+            blockNumberCached,
+          )}`,
         );
       }
       return blockNumberCached;
@@ -335,7 +341,7 @@ export class CommonService implements ICommonService {
         currentBlock,
         CommonService.latestBlockNumber,
         requestDetails,
-        this.ethBlockNumberCacheTtlMs
+        this.ethBlockNumberCacheTtlMs,
       );
 
       return currentBlock;
@@ -364,7 +370,7 @@ export class CommonService implements ICommonService {
   public async validateBlockHashAndAddTimestampToParams(
     params: any,
     blockHash: string,
-    requestDetails: RequestDetails
+    requestDetails: RequestDetails,
   ) {
     try {
       const block = await this.mirrorNodeClient.getBlock(blockHash, requestDetails);
@@ -397,7 +403,7 @@ export class CommonService implements ICommonService {
   public async getLogsByAddress(address: string | string[], params: any, requestDetails: RequestDetails) {
     const addresses = Array.isArray(address) ? address : [address];
     const logPromises = addresses.map((addr) =>
-      this.mirrorNodeClient.getContractResultsLogsByAddress(addr, requestDetails, params, undefined)
+      this.mirrorNodeClient.getContractResultsLogsByAddress(addr, requestDetails, params, undefined),
     );
 
     const logResults = await Promise.all(logPromises);
@@ -412,7 +418,7 @@ export class CommonService implements ICommonService {
   public async getLogsWithParams(
     address: string | string[] | null,
     params: any,
-    requestDetails: RequestDetails
+    requestDetails: RequestDetails,
   ): Promise<Log[]> {
     const EMPTY_RESPONSE = [];
 
@@ -439,8 +445,8 @@ export class CommonService implements ICommonService {
           removed: false,
           topics: log.topics,
           transactionHash: toHash32(log.transaction_hash),
-          transactionIndex: numberTo0x(log.transaction_index)
-        })
+          transactionIndex: numberTo0x(log.transaction_index),
+        }),
       );
     }
 
@@ -453,7 +459,7 @@ export class CommonService implements ICommonService {
     toBlock: string | 'latest',
     address: string | string[] | null,
     topics: any[] | null,
-    requestDetails: RequestDetails
+    requestDetails: RequestDetails,
   ): Promise<Log[]> {
     const EMPTY_RESPONSE = [];
     const params: any = {};
@@ -476,7 +482,7 @@ export class CommonService implements ICommonService {
   async resolveEvmAddress(
     address: string,
     requestDetails: RequestDetails,
-    searchableTypes = [constants.TYPE_CONTRACT, constants.TYPE_TOKEN, constants.TYPE_ACCOUNT]
+    searchableTypes = [constants.TYPE_CONTRACT, constants.TYPE_TOKEN, constants.TYPE_ACCOUNT],
   ): Promise<string> {
     if (!address) return address;
 
@@ -485,7 +491,7 @@ export class CommonService implements ICommonService {
       EthImpl.ethGetCode,
       requestDetails,
       searchableTypes,
-      0
+      0,
     );
     let resolvedAddress = address;
     if (
@@ -530,7 +536,7 @@ export class CommonService implements ICommonService {
       // for legacy EIP155 with tx.chainId=0x0, mirror-node will return a '0x' (EMPTY_HEX) value for contract result's chain_id
       //   which is incompatibile with certain tools (i.e. foundry). By setting this field, chainId, to undefined, the end jsonrpc
       //   object will leave out this field, which is the proper behavior for other tools to be compatible with.
-      chainId: cr.chain_id === '0x' ? undefined : cr.chain_id
+      chainId: cr.chain_id === '0x' ? undefined : cr.chain_id,
     };
 
     return TransactionFactory.createTransactionByType(cr.type, {
@@ -555,23 +561,23 @@ export class CommonService implements ICommonService {
     } catch (e: any) {
       this.logger.warn(
         e,
-        `${requestDetails.formattedRequestId} Mirror Node threw an error while retrieving fees. Fallback to consensus node.`
+        `${requestDetails.formattedRequestId} Mirror Node threw an error while retrieving fees. Fallback to consensus node.`,
       );
     }
 
     if (_.isNil(networkFees)) {
       if (this.logger.isLevelEnabled('debug')) {
         this.logger.debug(
-          `${requestDetails.formattedRequestId} Mirror Node returned no network fees. Fallback to consensus node.`
+          `${requestDetails.formattedRequestId} Mirror Node returned no network fees. Fallback to consensus node.`,
         );
       }
       networkFees = {
         fees: [
           {
             gas: await this.hapiService.getSDKClient().getTinyBarGasFee(callerName, requestDetails),
-            transaction_type: EthImpl.ethTxType
-          }
-        ]
+            transaction_type: EthImpl.ethTxType,
+          },
+        ],
       };
     }
 
@@ -603,7 +609,7 @@ export class CommonService implements ICommonService {
       let gasPrice: number | undefined = await this.cacheService.getAsync(
         constants.CACHE_KEY.GAS_PRICE,
         EthImpl.ethGasPrice,
-        requestDetails
+        requestDetails,
       );
 
       if (!gasPrice) {
@@ -614,7 +620,7 @@ export class CommonService implements ICommonService {
           gasPrice,
           EthImpl.ethGasPrice,
           requestDetails,
-          this.ethGasPriceCacheTtlMs
+          this.ethGasPriceCacheTtlMs,
         );
       }
 
@@ -639,5 +645,88 @@ export class CommonService implements ICommonService {
     } else {
       return Number(tag);
     }
+  }
+
+  private isBlockTagEarliest = (tag: string): boolean => {
+    return tag === EthImpl.blockEarliest;
+  };
+
+  private isBlockTagFinalized = (tag: string): boolean => {
+    return (
+      tag === EthImpl.blockFinalized ||
+      tag === EthImpl.blockLatest ||
+      tag === EthImpl.blockPending ||
+      tag === EthImpl.blockSafe
+    );
+  };
+
+  private isBlockNumValid = (num: string) => {
+    return /^0[xX]([1-9A-Fa-f]+[0-9A-Fa-f]{0,13}|0)$/.test(num) && Number.MAX_SAFE_INTEGER >= Number(num);
+  };
+
+  public isBlockParamValid = (tag: string | null) => {
+    return tag == null || this.isBlockTagEarliest(tag) || this.isBlockTagFinalized(tag) || this.isBlockNumValid(tag);
+  };
+
+  public isBlockHash = (blockHash: string): boolean => {
+    return new RegExp(constants.BLOCK_HASH_REGEX + '{64}$').test(blockHash);
+  };
+
+  /**
+   * Perform value format precheck before making contract call towards the mirror node
+   * @param {IContractCallRequest} transaction the transaction object
+   * @param {RequestDetails} requestDetails the request details for logging and tracking
+   */
+  public async contractCallFormat(transaction: IContractCallRequest, requestDetails: RequestDetails): Promise<void> {
+    if (transaction.value) {
+      transaction.value = weibarHexToTinyBarInt(transaction.value);
+    }
+    if (transaction.gasPrice) {
+      transaction.gasPrice = parseInt(transaction.gasPrice.toString());
+    } else {
+      transaction.gasPrice = await this.gasPrice(requestDetails).then((gasPrice) => parseInt(gasPrice));
+    }
+    if (transaction.gas) {
+      transaction.gas = parseInt(transaction.gas.toString());
+    }
+    if (!transaction.from && transaction.value && (transaction.value as number) > 0) {
+      if (ConfigService.get('OPERATOR_KEY_FORMAT') === 'HEX_ECDSA') {
+        transaction.from = this.hapiService.getMainClientInstance().operatorPublicKey?.toEvmAddress();
+      } else {
+        const operatorId = this.hapiService.getMainClientInstance().operatorAccountId!.toString();
+        const operatorAccount = await this.getAccount(operatorId, requestDetails);
+        transaction.from = operatorAccount?.evm_address;
+      }
+    }
+
+    // Support either data or input. https://ethereum.github.io/execution-apis/api-documentation/ lists input but many EVM tools still use data.
+    // We chose in the mirror node to use data field as the correct one, however for us to be able to support all tools,
+    // we have to modify transaction object, so that it complies with the mirror node.
+    // That means that, if input field is passed, but data is not, we have to copy value of input to the data to comply with mirror node.
+    // The second scenario occurs when both the data and input fields are present but hold different values.
+    // In this case, the value in the input field should be the one used for consensus based on this resource https://github.com/ethereum/execution-apis/blob/main/tests/eth_call/call-contract.io
+    // Eventually, for optimization purposes, we can rid of the input property or replace it with empty string.
+    if ((transaction.input && transaction.data === undefined) || (transaction.input && transaction.data)) {
+      transaction.data = transaction.input;
+      delete transaction.input;
+    }
+  }
+
+  /**
+   * Tries to get the account with the given address from the cache,
+   * if not found, it fetches it from the mirror node.
+   *
+   * @param {string} address the address of the account
+   * @param {RequestDetails} requestDetails the request details for logging and tracking
+   * @returns {Promise<IAccountInfo | null>} the account (if such exists for the given address)
+   */
+  public async getAccount(address: string, requestDetails: RequestDetails): Promise<IAccountInfo | null> {
+    const key = `${constants.CACHE_KEY.ACCOUNT}_${address}`;
+    let account = await this.cacheService.getAsync(key, EthImpl.ethEstimateGas, requestDetails);
+    if (!account) {
+      account = await this.mirrorNodeClient.getAccount(address, requestDetails);
+      await this.cacheService.set(key, account, EthImpl.ethEstimateGas, requestDetails);
+    }
+    return account;
   }
 }
