+++ packages/relay/src/lib/services/index.ts
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: Apache-2.0
+import { ContractService } from './contractService/ContractService';
+
+export { ContractService };
