+++ packages/relay/tests/lib/eth/eth_call.spec.ts
@@ -47,9 +47,10 @@ use(chaiAsPromised);
 let sdkClientStub: sinon.SinonStubbedInstance<SDKClient>;
 let getSdkClientStub: sinon.SinonStub;
 
-describe('@ethCall Eth Call spec', async function () {
+describe.only('@ethCall Eth Call spec', async function () {
   this.timeout(10000);
-  const { restMock, web3Mock, hapiServiceInstance, ethImpl, cacheService } = generateEthTestEnv();
+  const { restMock, web3Mock, hapiServiceInstance, ethImpl, cacheService, contractService, commonService } =
+    generateEthTestEnv();
 
   const ETH_CALL_REQ_ARGS = {
     from: ACCOUNT_ADDRESS_1,
@@ -92,8 +93,8 @@ describe('@ethCall Eth Call spec', async function () {
 
     beforeEach(() => {
       sandbox = sinon.createSandbox();
-      callConsensusNodeSpy = sandbox.spy(ethImpl, 'callConsensusNode');
-      callMirrorNodeSpy = sandbox.spy(ethImpl, 'callMirrorNode');
+      callConsensusNodeSpy = sandbox.spy(contractService, 'callConsensusNode');
+      callMirrorNodeSpy = sandbox.spy(contractService, 'callMirrorNode');
     });
 
     afterEach(() => {
@@ -132,12 +133,11 @@ describe('@ethCall Eth Call spec', async function () {
         );
         restMock.onGet(`contracts/${defaultCallData.to}`).reply(200, JSON.stringify(DEFAULT_CONTRACT));
 
-        await ethImpl.call(
+        await contractService.call(
           { ...defaultCallData, gas: `0x${defaultCallData.gas.toString(16)}` },
           'latest',
           requestDetails,
         );
-
         assert(callMirrorNodeSpy.calledOnce);
         assert(callConsensusNodeSpy.notCalled);
       });
@@ -156,7 +156,7 @@ describe('@ethCall Eth Call spec', async function () {
         );
         restMock.onGet(`contracts/${defaultCallData.to}`).reply(200, JSON.stringify(DEFAULT_CONTRACT));
 
-        await ethImpl.call(
+        await contractService.call(
           { ...defaultCallData, gas: `0x${defaultCallData.gas.toString(16)}` },
           'latest',
           requestDetails,
@@ -179,7 +179,7 @@ describe('@ethCall Eth Call spec', async function () {
         );
         restMock.onGet(`contracts/${defaultCallData.to}`).reply(200, JSON.stringify(DEFAULT_CONTRACT));
 
-        await ethImpl.call(
+        await contractService.call(
           { ...defaultCallData, gas: `0x${defaultCallData.gas.toString(16)}` },
           'latest',
           requestDetails,
@@ -197,7 +197,7 @@ describe('@ethCall Eth Call spec', async function () {
       web3Mock.onPost(`contracts/call`).reply(200, JSON.stringify({ result: '0x1' }));
 
       await expect(
-        ethImpl.call(
+        contractService.call(
           {
             from: ACCOUNT_ADDRESS_1,
             to: CONTRACT_ADDRESS_2,
@@ -218,7 +218,7 @@ describe('@ethCall Eth Call spec', async function () {
       web3Mock.onPost(`contracts/call`).reply(200, JSON.stringify({ result: '0x1' }));
 
       await expect(
-        ethImpl.call(
+        contractService.call(
           {
             from: ACCOUNT_ADDRESS_1,
             to: CONTRACT_ADDRESS_2,
@@ -244,8 +244,7 @@ describe('@ethCall Eth Call spec', async function () {
           return Uint8Array.of(0);
         },
       } as unknown as ContractFunctionResult);
-
-      const result = await ethImpl.call(
+      const result = await contractService.call(
         {
           from: ACCOUNT_ADDRESS_1,
           to: CONTRACT_ADDRESS_2,
@@ -254,7 +253,6 @@ describe('@ethCall Eth Call spec', async function () {
         'latest',
         requestDetails,
       );
-
       sinon.assert.calledWith(
         sdkClientStub.submitContractCallQueryWithRetry,
         CONTRACT_ADDRESS_2,
@@ -274,7 +272,7 @@ describe('@ethCall Eth Call spec', async function () {
         },
       } as unknown as ContractFunctionResult);
 
-      const result = await ethImpl.call(
+      const result = await contractService.call(
         {
           from: ACCOUNT_ADDRESS_1,
           to: CONTRACT_ADDRESS_2,
@@ -311,7 +309,7 @@ describe('@ethCall Eth Call spec', async function () {
         },
       } as unknown as ContractFunctionResult);
 
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.equal('0x00');
     });
 
@@ -324,7 +322,7 @@ describe('@ethCall Eth Call spec', async function () {
         gas: MAX_GAS_LIMIT,
       };
 
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
 
       expect((result as JsonRpcError).code).to.equal(-32014);
       expect((result as JsonRpcError).message).to.equal(
@@ -340,7 +338,7 @@ describe('@ethCall Eth Call spec', async function () {
         },
       } as unknown as ContractFunctionResult);
 
-      const result = await ethImpl.call(ETH_CALL_REQ_ARGS, 'latest', requestDetails);
+      const result = await contractService.call(ETH_CALL_REQ_ARGS, 'latest', requestDetails);
 
       sinon.assert.calledWith(
         sdkClientStub.submitContractCallQueryWithRetry,
@@ -363,7 +361,7 @@ describe('@ethCall Eth Call spec', async function () {
       } as unknown as ContractFunctionResult);
 
       for (let index = 0; index < 3; index++) {
-        const result = await ethImpl.call(
+        const result = await contractService.call(
           {
             from: ACCOUNT_ADDRESS_1,
             to: CONTRACT_ADDRESS_2,
@@ -381,7 +379,7 @@ describe('@ethCall Eth Call spec', async function () {
 
       const expectedError = predefined.INVALID_CONTRACT_ADDRESS(CONTRACT_ADDRESS_2);
       sdkClientStub.submitContractCallQueryWithRetry.throws(expectedError);
-      const call: string | JsonRpcError = await ethImpl.call(ETH_CALL_REQ_ARGS, 'latest', requestDetails);
+      const call: string | JsonRpcError = await contractService.call(ETH_CALL_REQ_ARGS, 'latest', requestDetails);
 
       expect((call as JsonRpcError).code).to.equal(expectedError.code);
       expect((call as JsonRpcError).message).to.equal(expectedError.message);
@@ -393,7 +391,7 @@ describe('@ethCall Eth Call spec', async function () {
         predefined.CONTRACT_REVERT(defaultErrorMessageText, defaultErrorMessageHex),
       );
 
-      const result = await ethImpl.call(ETH_CALL_REQ_ARGS, 'latest', requestDetails);
+      const result = await contractService.call(ETH_CALL_REQ_ARGS, 'latest', requestDetails);
 
       expect(result).to.exist;
       expect((result as JsonRpcError).code).to.equal(3);
@@ -427,7 +425,7 @@ describe('@ethCall Eth Call spec', async function () {
 
       sdkClientStub.submitContractCallQueryWithRetry.resolves(undefined);
 
-      const result = await ethImpl.call(
+      const result = await contractService.call(
         {
           to: CONTRACT_ADDRESS_2,
           data: CONTRACT_CALL_DATA,
@@ -484,7 +482,7 @@ describe('@ethCall Eth Call spec', async function () {
       restMock.onGet(`contracts/${CONTRACT_ADDRESS_2}`).reply(200, DEFAULT_CONTRACT_3_EMPTY_BYTECODE);
       web3Mock.onPost(`contracts/call`).replyOnce(200, {});
 
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.equal('0x');
     });
 
@@ -500,7 +498,7 @@ describe('@ethCall Eth Call spec', async function () {
 
       web3Mock.history.post = [];
 
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
 
       expect(web3Mock.history.post.length).to.gte(1);
       expect(web3Mock.history.post[0].data).to.equal(JSON.stringify({ ...callData, estimate: false, block: 'latest' }));
@@ -518,7 +516,7 @@ describe('@ethCall Eth Call spec', async function () {
       restMock.onGet(`contracts/${CONTRACT_ADDRESS_2}`).reply(200, JSON.stringify(DEFAULT_CONTRACT_2));
       await mockContractCall({ ...callData, block: 'latest' }, false, 200, { result: '0x00' }, requestDetails);
 
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.equal('0x00');
     });
 
@@ -530,7 +528,7 @@ describe('@ethCall Eth Call spec', async function () {
         gas: MAX_GAS_LIMIT,
       };
       await mockContractCall({ ...callData, block: 'latest' }, false, 200, { result: '0x00' }, requestDetails);
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.equal('0x00');
     });
 
@@ -543,7 +541,7 @@ describe('@ethCall Eth Call spec', async function () {
         gas: MAX_GAS_LIMIT,
       };
       await mockContractCall({ ...callData, block: 'latest' }, false, 200, { result: '0x00' }, requestDetails);
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.equal('0x00');
     });
 
@@ -561,7 +559,7 @@ describe('@ethCall Eth Call spec', async function () {
         },
         requestDetails,
       );
-      const res = await ethImpl.call(callData, 'latest', requestDetails);
+      const res = await contractService.call(callData, 'latest', requestDetails);
       expect(res).to.equal('0x00');
     });
 
@@ -580,7 +578,11 @@ describe('@ethCall Eth Call spec', async function () {
       restMock.onGet(`contracts/${CONTRACT_ADDRESS_2}`).reply(200, JSON.stringify(DEFAULT_CONTRACT_2));
 
       // Relay is called with value in Weibars
-      const result = await ethImpl.call({ ...callData, value: ONE_TINYBAR_IN_WEI_HEX }, 'latest', requestDetails);
+      const result = await contractService.call(
+        { ...callData, value: ONE_TINYBAR_IN_WEI_HEX },
+        'latest',
+        requestDetails,
+      );
       expect(result).to.equal('0x00');
     });
 
@@ -593,7 +595,7 @@ describe('@ethCall Eth Call spec', async function () {
         gas: MAX_GAS_LIMIT,
       };
       await mockContractCall({ ...callData, block: 'latest' }, false, 429, mockData.tooManyRequests, requestDetails);
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.be.not.null;
       expect((result as JsonRpcError).code).to.eq(-32605);
     });
@@ -608,7 +610,7 @@ describe('@ethCall Eth Call spec', async function () {
       };
       restMock.onGet(`contracts/${CONTRACT_ADDRESS_2}`).reply(200, JSON.stringify(DEFAULT_CONTRACT_2));
       await mockContractCall({ ...callData, block: 'latest' }, false, 400, mockData.contractReverted, requestDetails);
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.be.not.null;
       expect((result as JsonRpcError).code).to.eq(3);
       expect((result as JsonRpcError).message).to.contain(mockData.contractReverted._status.messages[0].message);
@@ -632,8 +634,7 @@ describe('@ethCall Eth Call spec', async function () {
         },
       } as unknown as ContractFunctionResult);
 
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
-
+      const result = await contractService.call(callData, 'latest', requestDetails);
       sinon.assert.calledWith(
         sdkClientStub.submitContractCallQueryWithRetry,
         CONTRACT_ADDRESS_2,
@@ -657,7 +658,7 @@ describe('@ethCall Eth Call spec', async function () {
       restMock.onGet(`contracts/${CONTRACT_ADDRESS_2}`).reply(200, JSON.stringify(DEFAULT_CONTRACT_2));
       await mockContractCall({ ...callData, block: 'latest' }, false, 400, mockData.contractReverted, requestDetails);
       sinon.reset();
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       sinon.assert.notCalled(sdkClientStub.submitContractCallQueryWithRetry);
       expect(result).to.not.be.null;
       expect((result as JsonRpcError).code).to.eq(3);
@@ -692,7 +693,7 @@ describe('@ethCall Eth Call spec', async function () {
         requestDetails,
       );
 
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
 
       expect(result).to.exist;
       expect((result as JsonRpcError).code).to.eq(3);
@@ -732,7 +733,7 @@ describe('@ethCall Eth Call spec', async function () {
       };
 
       await mockContractCall({ ...callData, block: 'latest' }, false, 400, mockData.invalidTransaction, requestDetails);
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.be.not.null;
       expect(result).to.equal('0x');
     });
@@ -747,7 +748,7 @@ describe('@ethCall Eth Call spec', async function () {
       };
 
       await mockContractCall({ ...callData, block: 'latest' }, false, 400, mockData.failInvalid, requestDetails);
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.be.not.null;
       expect(result).to.equal('0x');
     });
@@ -766,7 +767,7 @@ describe('@ethCall Eth Call spec', async function () {
         { result: EXAMPLE_CONTRACT_BYTECODE },
         requestDetails,
       );
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.eq(EXAMPLE_CONTRACT_BYTECODE);
     });
 
@@ -783,7 +784,7 @@ describe('@ethCall Eth Call spec', async function () {
         { result: EXAMPLE_CONTRACT_BYTECODE },
         requestDetails,
       );
-      const result = await ethImpl.call(callData, 'latest', requestDetails);
+      const result = await contractService.call(callData, 'latest', requestDetails);
       expect(result).to.eq(EXAMPLE_CONTRACT_BYTECODE);
     });
 
@@ -818,7 +819,7 @@ describe('@ethCall Eth Call spec', async function () {
       requestDetails: RequestDetails,
     ) {
       const formattedCallData = { ...callData, estimate };
-      await ethImpl.contractCallFormat(formattedCallData, requestDetails);
+      await contractService.contractCallFormat(formattedCallData, requestDetails);
       return web3Mock.onPost('contracts/call', formattedCallData).reply(statusCode, JSON.stringify(result));
     }
   });
@@ -842,7 +843,7 @@ describe('@ethCall Eth Call spec', async function () {
         value: '0x2540BE400',
       };
 
-      await ethImpl.contractCallFormat(transaction, requestDetails);
+      await contractService.contractCallFormat(transaction, requestDetails);
       expect(transaction.value).to.equal(1);
     });
 
@@ -851,7 +852,7 @@ describe('@ethCall Eth Call spec', async function () {
         gasPrice: '1000000000',
       };
 
-      await ethImpl.contractCallFormat(transaction, requestDetails);
+      await contractService.contractCallFormat(transaction, requestDetails);
 
       expect(transaction.gasPrice).to.equal(1000000000);
     });
@@ -861,7 +862,7 @@ describe('@ethCall Eth Call spec', async function () {
         gas: '50000',
       };
 
-      await ethImpl.contractCallFormat(transaction, requestDetails);
+      await contractService.contractCallFormat(transaction, requestDetails);
 
       expect(transaction.gas).to.equal(50000);
     });
@@ -873,7 +874,7 @@ describe('@ethCall Eth Call spec', async function () {
         input: inputValue,
         data: dataValue,
       };
-      await ethImpl.contractCallFormat(transaction, requestDetails);
+      await contractService.contractCallFormat(transaction, requestDetails);
       expect(transaction.data).to.eq(inputValue);
       expect(transaction.data).to.not.eq(dataValue);
       expect(transaction.input).to.be.undefined;
@@ -884,7 +885,7 @@ describe('@ethCall Eth Call spec', async function () {
       const transaction = {
         data: dataValue,
       };
-      await ethImpl.contractCallFormat(transaction, requestDetails);
+      await contractService.contractCallFormat(transaction, requestDetails);
       expect(transaction.data).to.eq(dataValue);
     });
 
@@ -893,7 +894,7 @@ describe('@ethCall Eth Call spec', async function () {
         input: 'input data',
       };
 
-      await ethImpl.contractCallFormat(transaction, requestDetails);
+      await contractService.contractCallFormat(transaction, requestDetails);
 
       // @ts-ignore
       expect(transaction.data).to.equal('input data');
@@ -907,7 +908,7 @@ describe('@ethCall Eth Call spec', async function () {
         gas: '50000',
       };
 
-      await ethImpl.contractCallFormat(transaction, requestDetails);
+      await contractService.contractCallFormat(transaction, requestDetails);
 
       expect(transaction.value).to.equal(1);
       expect(transaction.gasPrice).to.equal(1000000000);
@@ -920,9 +921,9 @@ describe('@ethCall Eth Call spec', async function () {
         gasPrice: undefined,
       };
 
-      await ethImpl.contractCallFormat(transaction, requestDetails);
+      await contractService.contractCallFormat(transaction, requestDetails);
 
-      const expectedGasPrice = await ethImpl.gasPrice(requestDetails);
+      const expectedGasPrice = await commonService.gasPrice(requestDetails);
       expect(transaction.gasPrice).to.equal(parseInt(expectedGasPrice));
     });
 
@@ -933,7 +934,7 @@ describe('@ethCall Eth Call spec', async function () {
         from: undefined,
       };
 
-      await ethImpl.contractCallFormat(transaction, requestDetails);
+      await contractService.contractCallFormat(transaction, requestDetails);
 
       expect(transaction.from).to.equal(operatorEvmAddress);
     });
@@ -953,8 +954,8 @@ describe('@ethCall Eth Call spec', async function () {
 
     beforeEach(() => {
       sandbox = sinon.createSandbox();
-      callConsensusNodeSpy = sandbox.spy(ethImpl, 'callConsensusNode');
-      callMirrorNodeSpy = sandbox.spy(ethImpl, 'callMirrorNode');
+      callConsensusNodeSpy = sandbox.spy(contractService, 'callConsensusNode');
+      callMirrorNodeSpy = sandbox.spy(contractService, 'callMirrorNode');
     });
 
     afterEach(() => {
@@ -962,7 +963,7 @@ describe('@ethCall Eth Call spec', async function () {
     });
 
     it('eth_call with matched selector redirects to consensus', async function () {
-      await ethImpl.call(
+      await contractService.call(
         {
           to: ACCOUNT_ADDRESS_1,
           data: REDIRECTED_SELECTOR,
@@ -976,7 +977,7 @@ describe('@ethCall Eth Call spec', async function () {
     });
 
     it('eth_call with non-matched selector redirects to consensus', async function () {
-      await ethImpl.call(
+      await contractService.call(
         {
           to: ACCOUNT_ADDRESS_1,
           data: NON_REDIRECTED_SELECTOR,
