+++ packages/relay/tests/lib/eth/eth_estimateGas.spec.ts
@@ -31,8 +31,17 @@ let ethImplOverridden: Eth;
 
 describe('@ethEstimateGas Estimate Gas spec', async function () {
   this.timeout(10000);
-  const { restMock, web3Mock, hapiServiceInstance, ethImpl, cacheService, mirrorNodeInstance, logger, registry } =
-    generateEthTestEnv();
+  const {
+    restMock,
+    web3Mock,
+    hapiServiceInstance,
+    ethImpl,
+    cacheService,
+    mirrorNodeInstance,
+    logger,
+    registry,
+    contractService,
+  } = generateEthTestEnv();
 
   const requestDetails = new RequestDetails({ requestId: 'eth_estimateGasTest', ipAddress: '0.0.0.0' });
 
@@ -44,12 +53,14 @@ describe('@ethEstimateGas Estimate Gas spec', async function () {
     requestDetails: RequestDetails,
   ) {
     const formattedData = { ...callData, estimate };
-    await ethImpl.contractCallFormat(formattedData, requestDetails);
+    await contractService.contractCallFormat(formattedData, requestDetails);
     return web3Mock.onPost('contracts/call', formattedData).reply(statusCode, JSON.stringify(result));
   }
 
   function mockGetAccount(idOrAliasOrEvmAddress: string, statusCode: number, result: any) {
-    return restMock.onGet(`accounts/${idOrAliasOrEvmAddress}?transactions=false`).reply(statusCode, JSON.stringify(result));
+    return restMock
+      .onGet(`accounts/${idOrAliasOrEvmAddress}?transactions=false`)
+      .reply(statusCode, JSON.stringify(result));
   }
 
   const transaction = {
@@ -181,7 +192,9 @@ describe('@ethEstimateGas Estimate Gas spec', async function () {
       value: 10, //in tinybars
     };
     await mockContractCall(callData, true, 501, { errorMessage: '', statusCode: 501 }, requestDetails);
-    restMock.onGet(`accounts/${RECEIVER_ADDRESS}${NO_TRANSACTIONS}`).reply(200, JSON.stringify({ address: RECEIVER_ADDRESS }));
+    restMock
+      .onGet(`accounts/${RECEIVER_ADDRESS}${NO_TRANSACTIONS}`)
+      .reply(200, JSON.stringify({ address: RECEIVER_ADDRESS }));
 
     const gas = await ethImpl.estimateGas(
       {
@@ -200,7 +213,9 @@ describe('@ethEstimateGas Estimate Gas spec', async function () {
       value: 10, //in tinybars
     };
     await mockContractCall(callData, true, 501, { errorMessage: '', statusCode: 501 }, requestDetails);
-    restMock.onGet(`accounts/${RECEIVER_ADDRESS}${NO_TRANSACTIONS}`).reply(200, JSON.stringify({ address: RECEIVER_ADDRESS }));
+    restMock
+      .onGet(`accounts/${RECEIVER_ADDRESS}${NO_TRANSACTIONS}`)
+      .reply(200, JSON.stringify({ address: RECEIVER_ADDRESS }));
 
     const gasBeforeCache = await ethImpl.estimateGas(
       {
@@ -251,7 +266,9 @@ describe('@ethEstimateGas Estimate Gas spec', async function () {
       value: 0, //in tinybars
     };
     await mockContractCall(callData, true, 501, { errorMessage: '', statusCode: 501 }, requestDetails);
-    restMock.onGet(`accounts/${RECEIVER_ADDRESS}${NO_TRANSACTIONS}`).reply(200, JSON.stringify({ address: RECEIVER_ADDRESS }));
+    restMock
+      .onGet(`accounts/${RECEIVER_ADDRESS}${NO_TRANSACTIONS}`)
+      .reply(200, JSON.stringify({ address: RECEIVER_ADDRESS }));
     const gas = await ethImpl.estimateGas(
       {
         to: RECEIVER_ADDRESS,
@@ -287,7 +304,9 @@ describe('@ethEstimateGas Estimate Gas spec', async function () {
       value: -100_000_000_000, //in tinybars
     };
     await mockContractCall(callData, true, 501, { errorMessage: '', statusCode: 501 }, requestDetails);
-    restMock.onGet(`accounts/${RECEIVER_ADDRESS}${NO_TRANSACTIONS}`).reply(200, JSON.stringify({ address: RECEIVER_ADDRESS }));
+    restMock
+      .onGet(`accounts/${RECEIVER_ADDRESS}${NO_TRANSACTIONS}`)
+      .reply(200, JSON.stringify({ address: RECEIVER_ADDRESS }));
     const result = await ethImpl.estimateGas(
       {
         to: RECEIVER_ADDRESS,
@@ -530,7 +549,7 @@ describe('@ethEstimateGas Estimate Gas spec', async function () {
       gas: '0xd97010',
     };
 
-    await ethImpl.contractCallFormat(transaction, requestDetails);
+    await contractService.contractCallFormat(transaction, requestDetails);
     expect(transaction.value).to.eq(1110);
     expect(transaction.gasPrice).to.eq(1000000);
     expect(transaction.gas).to.eq(14250000);
@@ -548,7 +567,7 @@ describe('@ethEstimateGas Estimate Gas spec', async function () {
       gas: '0xd97010',
     };
 
-    await ethImpl.contractCallFormat(transaction, requestDetails);
+    await contractService.contractCallFormat(transaction, requestDetails);
     expect(transaction.data).to.eq(inputValue);
     expect(transaction.data).to.not.eq(dataValue);
     expect(transaction.input).to.be.undefined;
