+++ packages/relay/tests/lib/eth/eth_getBlockReceipts.spec.ts
@@ -9,6 +9,7 @@ import { numberTo0x } from '../../../dist/formatters';
 import { SDKClient } from '../../../src/lib/clients';
 import constants from '../../../src/lib/constants';
 import { EthImpl } from '../../../src/lib/eth';
+import { ContractService } from '../../../src/lib/services';
 import { CacheService } from '../../../src/lib/services/cacheService/cacheService';
 import HAPIService from '../../../src/lib/services/hapiService/hapiService';
 import { RequestDetails } from '../../../src/lib/types';
@@ -40,11 +41,13 @@ describe('@ethGetBlockReceipts using MirrorNode', async function () {
     restMock,
     hapiServiceInstance,
     ethImpl,
+    contractService,
     cacheService,
   }: {
     restMock: MockAdapter;
     hapiServiceInstance: HAPIService;
     ethImpl: EthImpl;
+    contractService: ContractService;
     cacheService: CacheService;
   } = generateEthTestEnv(true);
   const results = defaultContractResults.results;
@@ -54,7 +57,9 @@ describe('@ethGetBlockReceipts using MirrorNode', async function () {
     // reset cache and restMock
     await cacheService.clear(requestDetails);
     currentGasPriceStub = sinon.stub(ethImpl, 'getCurrentGasPriceForBlock').resolves('0x25');
-    extractBlockNumberOrTagStub = sinon.stub(ethImpl, 'extractBlockNumberOrTag').resolves(BLOCK_NUMBER);
+    extractBlockNumberOrTagStub = sinon
+      .stub(contractService, 'extractBlockNumberOrTag')
+      .resolves(BLOCK_NUMBER.toString());
     sdkClientStub = sinon.createStubInstance(SDKClient);
     getSdkClientStub = sinon.stub(hapiServiceInstance, 'getSDKClient').returns(sdkClientStub);
     restMock.onGet('network/fees').reply(200, JSON.stringify(DEFAULT_NETWORK_FEES));
